Generate more realistic polygon object
