+++ src/helpers/zoneShapePropsFor.ts
@@ -1,7 +1,9 @@
-import { CIRCLE_SHAPE, CIRCULAR_RING_SHAPE, CIRCULAR_SECTOR_SHAPE, POLYGON_SHAPE, RECTANGLE_SHAPE, SECTION_CIRCULAR_RING_SHAPE } from '../constants'
+import { Feature, Polygon, Position } from 'geojson'
+import { CIRCLE_SHAPE, CIRCULAR_RING_SHAPE, CIRCULAR_SECTOR_SHAPE, POLYGON_SHAPE, RECTANGLE_SHAPE, SECTION_CIRCULAR_RING_SHAPE, ZONE_TYPE } from '../constants'
+import { ZoneProps } from '../types'
 import { ZoneRectangleProps, ZoneShapeProps } from '../zoneShapeTypes'
 
-export const zoneShapePropsFor = (shapeType: string): ZoneShapeProps => {
+const zoneShapePropsFor = (shapeType: string): ZoneShapeProps => {
   switch (shapeType) {
   case RECTANGLE_SHAPE:
     return {
@@ -11,8 +13,7 @@ export const zoneShapePropsFor = (shapeType: string): ZoneShapeProps => {
     } as ZoneRectangleProps  
   case POLYGON_SHAPE:
     return {
-      shapeType: POLYGON_SHAPE,
-      points: [[22, 33], [33, 44]],
+      shapeType: POLYGON_SHAPE
     }
   case CIRCULAR_RING_SHAPE:
     return {
@@ -43,4 +44,24 @@ export const zoneShapePropsFor = (shapeType: string): ZoneShapeProps => {
   default:
     throw new Error(`Unknown shape type: ${shapeType}`)  
   }
+}
+
+export const zoneFeatureFor = (shapeType: string) => {
+  const isPolygon = shapeType === POLYGON_SHAPE
+  const coords: Position[] = isPolygon ? [[22, 23],[33, 34], [34, 33],[22, 23]] : []
+  const zone: Feature<Polygon, ZoneProps> = {
+    type: 'Feature',
+    properties: {
+      name: '',
+      dataType: ZONE_TYPE,
+      specifics:  zoneShapePropsFor(shapeType),
+      color: '#FF0000',
+      visible: true,
+    },
+    geometry: {
+      type: 'Polygon',
+      coordinates: [coords],
+    },
+  }
+  return zone
 }
